Updated API document. JS - Fixed issue issue in MDSClient. JS - Fixed issue in Associations. JS - Updated Cache to create BIGINT column to store larger values than INT. JS - Published as 2.0.53.

+++ Source/JS/src/Server/Cache.js
@@ -31,7 +31,7 @@ module.exports = class Cache {
 											// // Check value type
 											// if (valueInfo.valueType == 'integer')
 												// Integer
-												return new TableColumn.INT(valueInfo.name,
+												return new TableColumn.BIGINT(valueInfo.name,
 														TableColumn.options.nonNull | TableColumn.options.unsigned);
 											}));
 		this.table = statementPerformer.table(tableName, tableColumns);
